requirements/tests.txt
pytest-aiohttp>=0.3.0
